organize component proj - MVVM
