+++ libraries/lokqlDxComponents/Models/ShowCompletionOptions.cs
@@ -1,0 +1,14 @@
+using AvaloniaEdit.CodeCompletion;
+using Intellisense;
+
+namespace lokqlDxComponents.Models;
+
+public readonly record struct ShowCompletionOptions()
+{
+    public IReadOnlyCollection<IntellisenseEntry> Completions { get; init; } = [];
+    public string Prefix { get; init; } = string.Empty;
+    public int Rewind { get; init; }
+    public Action<CompletionWindow> OnCompletionWindowDataPopulated { get; init; } = EmptyAction;
+    private static readonly Action<CompletionWindow> EmptyAction = _ => { };
+    public static ShowCompletionOptions Empty => new();
+}
